scp-2709 - ACTUS code refactoring
* moved test data structures into test directory
