+++ marlowe-actus/src/Language/Marlowe/ACTUS/Definitions/BusinessEvents.hs
@@ -5,8 +5,6 @@
 module Language.Marlowe.ACTUS.Definitions.BusinessEvents where
 
 import           Data.Aeson.Types (ToJSON)
-import           Data.Map
-import           Data.Time
 import           GHC.Generics     (Generic)
 
 {-| ACTUS event types
@@ -49,18 +47,3 @@ data RiskFactors = RiskFactors
     deriving stock (Generic)
     deriving (Show, ToJSON)
 
-{-| Observed data
--}
-type DataObserved = Map String ValuesObserved
-
-data ValuesObserved = ValuesObserved
-  { identifier :: String
-  , values     :: [ValueObserved]
-  }
-  deriving (Show)
-
-data ValueObserved = ValueObserved
-  { timestamp :: Day
-  , value     :: Double
-  }
-  deriving (Show)
